packages/vca_auth/pyproject.toml
[project]
name = "vca-auth"
version = "0.1.0"
requires-python = ">=3.11,<3.12"
dependencies = ["sqlmodel", "pydantic", "pydantic-settings", "python-ulid>=3.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vca_auth"]
